# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  service.AICoreConfigurationRequest:
    properties:
      executableId:
        type: string
      inputArtifactBindings:
        items:
          additionalProperties:
            type: string
          type: object
        type: array
      name:
        type: string
      parameterBindings:
        description: model name and model version
        items:
          additionalProperties:
            type: string
          type: object
        type: array
      scenarioId:
        type: string
    required:
    - executableId
    - name
    - scenarioId
    type: object
  service.AICoreConfigurationResponse:
    properties:
      id:
        type: string
      message:
        type: string
    type: object
  service.AICoreConfigurationsResponse:
    properties:
      count:
        type: integer
      resources:
        items:
          $ref: '#/definitions/service.AICoreConfiguration'
        type: array
    type: object
...
  service.AICoreDeploymentDeletionResponse:
    properties:
      id:
        type: string
      message:
        type: string
    type: object
...
  service.AICoreDeploymentModificationRequest:
    properties:
      configurationId:
        type: string
      targetStatus:
        type: string
    type: object
...
  service.AICoreDeploymentRequest:
    properties:
      configurationId:
        type: string
      configurationRequest:
        $ref: '#/definitions/service.AICoreConfigurationRequest'
      ttl:
        type: string
    type: object
  service.AICoreDeploymentResponse:
    properties:
      deploymentUrl:
        type: string
      id:
        type: string
      message:
        type: string
      status:
        type: string
      ttl:
        type: string
    type: object
  service.AICoreDeploymentsResponse:
    properties:
      count:
        type: integer
      deployments:
        items:
          $ref: '#/definitions/service.AICoreTeamDeployments'
        type: array
    type: object
...
  service.AICoreInferenceRequest:
    properties:
      deploymentId:
        type: string
      max_tokens:
        type: integer
      messages:
        items:
          $ref: '#/definitions/service.AICoreInferenceMessage'
        minItems: 1
        type: array
      stream:
        type: boolean
      temperature:
        type: number
      top_p:
        type: number
    required:
    - deploymentId
    - messages
    type: object
  service.AICoreInferenceResponse:
    properties:
      choices:
        items:
          $ref: '#/definitions/service.AICoreInferenceChoice'
        type: array
      created:
        type: integer
      id:
        type: string
      model:
        type: string
      object:
        type: string
      system_fingerprint:
        type: string
      usage:
        $ref: '#/definitions/service.AICoreInferenceUsage'
    type: object
...
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: This is the backend API for the Developer Portal, providing endpoints
    for managing organizations, teams, projects, components, and deployments.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Developer Portal Backend API
  version: "1.0"
paths:
  /ai-core/chat/inference:
    post:
      consumes:
      - application/json
      description: Send chat messages to a deployed model and get AI responses (supports
        streaming)
      parameters:
      - description: Chat inference request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.AICoreInferenceRequest'
      produces:
      - application/json
      - text/event-stream
      responses:
        "200":
          description: Successfully received inference response
          schema:
            $ref: '#/definitions/service.AICoreInferenceResponse'
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: User not assigned to team or team credentials not found
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Deployment not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Chat inference with AI Core deployed model
      tags:
      - ai-core
  /ai-core/configurations:
    get:
      consumes:
      - application/json
      description: Get all configurations from AI Core for the authenticated user's
        team
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved configurations
          schema:
            $ref: '#/definitions/service.AICoreConfigurationsResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: User not assigned to team or team credentials not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get AI Core configurations
      tags:
      - ai-core
    post:
      consumes:
      - application/json
      description: Create a new configuration in AI Core for the authenticated user's
        team
      parameters:
      - description: Configuration data
        in: body
        name: configuration
        required: true
        schema:
          $ref: '#/definitions/service.AICoreConfigurationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created configuration
          schema:
            $ref: '#/definitions/service.AICoreConfigurationResponse'
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: User not assigned to team or team credentials not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create AI Core configuration
      tags:
      - ai-core
  /ai-core/deployments:
    get:
      consumes:
      - application/json
      description: Get all deployments from AI Core for the authenticated user's team
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved deployments
          schema:
            $ref: '#/definitions/service.AICoreDeploymentsResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: User not assigned to team or team credentials not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get AI Core deployments
      tags:
      - ai-core
    post:
      consumes:
      - application/json
      description: Create a new deployment in AI Core using either an existing configuration
        ID or by creating a new configuration first
      parameters:
      - description: Deployment data - either configurationId or configurationRequest
          must be provided
        in: body
        name: deployment
        required: true
        schema:
          $ref: '#/definitions/service.AICoreDeploymentRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Successfully scheduled deployment
          schema:
            $ref: '#/definitions/service.AICoreDeploymentResponse'
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: User not assigned to team or team credentials not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create AI Core deployment
      tags:
      - ai-core
  /ai-core/deployments/{deploymentId}:
    delete:
      consumes:
      - application/json
      description: Mark deployment as deleted in AI Core
      parameters:
      - description: Deployment ID
        in: path
        name: deploymentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Successfully scheduled deployment deletion
          schema:
            $ref: '#/definitions/service.AICoreDeploymentDeletionResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: User not assigned to team or team credentials not found
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Deployment not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete AI Core deployment
      tags:
      - ai-core
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific deployment from AI Core
      parameters:
      - description: Deployment ID
        in: path
        name: deploymentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved deployment details
          schema:
            $ref: '#/definitions/service.AICoreDeploymentDetailsResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: User not assigned to team or team credentials not found
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Deployment not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get AI Core deployment details
      tags:
      - ai-core
    patch:
      consumes:
      - application/json
      description: Update target status or configuration of a deployment in AI Core
      parameters:
      - description: Deployment ID
        in: path
        name: deploymentId
        required: true
        type: string
      - description: Deployment modification data
        in: body
        name: modification
        required: true
        schema:
          $ref: '#/definitions/service.AICoreDeploymentModificationRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Successfully scheduled deployment modification
          schema:
            $ref: '#/definitions/service.AICoreDeploymentModificationResponse'
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: User not assigned to team or team credentials not found
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Deployment not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update AI Core deployment
      tags:
      - ai-core
  /ai-core/me:
    get:
      consumes:
      - application/json
      description: Returns current username and AI instances derived from metadata
        and team role
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved AI Core user context
          schema:
            $ref: '#/definitions/service.AICoreMeResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get AI Core user context
      tags:
      - ai-core
  /ai-core/models:
    get:
      consumes:
      - application/json
      description: Get all available models from AI Core for a specific scenario
      parameters:
      - description: Scenario ID to get models for
        in: query
        name: scenarioId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved models
          schema:
            $ref: '#/definitions/service.AICoreModelsResponse'
        "400":
          description: Bad request - missing scenarioId parameter
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: User not assigned to team or team credentials not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get AI Core models
      tags:
      - ai-core
  /ai-core/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload multiple files for use in multimodal AI requests (combined
        max 5MB)
      parameters:
      - description: Files to upload (multiple allowed)
        in: formData
        name: files
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Successfully uploaded files
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid files or size
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Upload file attachments for AI inference
      tags:
      - ai-core
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"